third_party/WebKit/LayoutTests/fast/css/namespaces/namespaces-invalid-prefix.xhtml

<html xmlns="http://www.w3.org/1999/xhtml">
<style>
abc|div {
    background-color: red;
}

abc|div#target {
    background-color: red;
}

[abc|id] {
    background-color: green;
}
</style>
<div id="target"></div>

<script src="../../../resources/testharness.js"></script>
<script src="../../../resources/testharnessreport.js"></script>

<script>
test(function(){
    assert_equals(document.styleSheets[0].cssRules.length, 0);
}, "Style rules with namespace prefixes not declared before, should be ignored");
</script>

</html>
